refactor: game 패키지 operation으로 변경
game 패키지 operation으로 변경

- 게임 운영과 관련된 도메인을 operation 패키지로 분리
